Fix the problem with context menu not going away.
When you right click the notification icon, a context
menu appears, clicking outside of this menu should
cancel this menu. But without this change the menu
would stick around.
From:
https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-trackpopupmenu
"To display a context menu for a notification icon, the current window
must be the foreground window before the application calls
TrackPopupMenu or TrackPopupMenuEx. Otherwise, the menu will not
disappear when the user clicks outside of the menu or the window that
created the menu (if it is visible)."